clean up [leagueId].tsx
separated util funcs and Scoreboard component to separate files
