+++ frontend/src/app/(app)/leagues/[leagueId].tsx
@@ -1,134 +1,20 @@
-import { BackHandler, ScrollView, StyleSheet, View } from 'react-native';
+import { BackHandler, StyleSheet, View } from 'react-native';
 import React, { useEffect, useState } from 'react';
 import { Stack, useLocalSearchParams, useNavigation, useRouter } from 'expo-router';
 
-import { colors, globalStyles, scoreboardWidths } from '../../../styles/styles';
-// import { fetchLeagueById } from '../../../utils/fetchLeague';
-// import { SelectedLeague } from '../../../types/League';
+import { globalStyles } from '../../../styles/styles';
 import PremText from '../../../components/basic/PremText';
-import PlayerScore from '../../../components/leagueId/PlayerScore';
-import { Player, PlayerPoints, ScoreboardPlayer } from '../../../types/Player';
 import PremButton from '../../../components/basic/PremButton';
 import GameweekShifter from '../../../components/basic/GameweekShifter';
 import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
 import { getSelectedLeague, unselect } from '../../../redux/reducers/leaguesReducer';
-import User from '../../../types/User';
-import Gameweek from '../../../types/Gameweek';
-
-const calculatePoints = (points: PlayerPoints[], gw: number) => {
-  return points.reduce((sum, curr) => {
-    if (curr.gw > gw) return sum;
-    return (sum += curr.points);
-  }, 0);
-};
-
-const getScoreboardedPlayers = (players: Player[], selectedGw: number): ScoreboardPlayer[] => {
-  const scoreboardPlayers = players.map((player) => {
-    const playerPoints = calculatePoints(player.points, selectedGw);
-    const playerPrevPoints = calculatePoints(player.points, selectedGw - 1);
-
-    return { id: player.id, name: player.name, points: playerPoints, prevPoints: playerPrevPoints };
-  });
-
-  const copy = scoreboardPlayers.slice();
-
-  const finalPlayers = scoreboardPlayers.map((player) => {
-    const prevPos =
-      copy
-        .sort((a, b) => (a.prevPoints >= b.prevPoints ? -1 : 1))
-        .findIndex((p) => p.id === player.id) + 1;
-
-    const currPos =
-      copy.sort((a, b) => (a.points >= b.points ? -1 : 1)).findIndex((p) => p.id === player.id) + 1;
-
-    return { ...player, position: currPos, prevPosition: prevPos, posChange: prevPos - currPos };
-  });
-
-  return finalPlayers.sort((a, b) => (a.points >= b.points ? -1 : 1));
-};
-
-const calculateTimeUntilGW = (gameweek: Gameweek) => {
-  const now = new Date();
-  const opens = new Date(gameweek.opens);
-  const closes = new Date(gameweek.closes);
-  const finishes = new Date(gameweek.finishes);
-
-  // "now" can be inbetween any of these 3 times, giving us four cases
-  // opens - closes - finishes
-  if (now < opens)
-    return `Opens on ${opens.toDateString()}, ${opens.getHours()}:${opens.getMinutes()}`;
-
-  if (finishes < now)
-    return `Finished on ${finishes.toDateString()}, ${finishes.getHours()}:${finishes.getMinutes()}`;
-
-  if (now < closes) return `Closes at ${closes.toDateString()}`;
-
-  return `Closed on ${closes.toDateString()}, ${closes.getHours()}:${closes.getMinutes()}`;
-};
-
-const Scoreboard = (players: User[], gw: number, myId: string) => {
-  const playerItems = players.map((player, index) => (
-    <PlayerScore
-      position={index + 1}
-      player={player}
-      userId={myId}
-      key={player.id}
-      gw={gw}
-      leagueSize={players.length}
-    />
-  ));
-
-  return (
-    <View style={styles.scoreboard}>
-      <View style={styles.scoreboardHeader}>
-        <PremText centered>Scoreboard</PremText>
-      </View>
-      <View style={styles.header}>
-        <View style={[styles.textWrapper]}>
-          <View
-            style={[
-              styles.textWrapper,
-              // globalStyles.border,
-              players.length >= 10
-                ? scoreboardWidths.between10and100WrapperWidth
-                : scoreboardWidths.under10WrapperWidth,
-            ]}
-          >
-            <PremText order={4}>pos</PremText>
-            <PremText order={4}>+/-</PremText>
-          </View>
-          <PremText order={4}>name</PremText>
-        </View>
-        <View style={[styles.textWrapper, scoreboardWidths.pointsWidth]}>
-          <PremText order={4}>{`gw${gw}`}</PremText>
-          <PremText order={4}>points</PremText>
-        </View>
-      </View>
-      <ScrollView style={{ maxHeight: 400 }}>
-        <View style={styles.scoreboardScrollWrapper}>{playerItems}</View>
-      </ScrollView>
-    </View>
-  );
-};
-
-// const calculateYourPlace = (players: ScoreboardPlayer[], userId: string) => {
-//   return players.findIndex((player) => player.id === userId) + 1;
-// };
-const calculateYourPlace = (players: User[], userId: string) => {
-  return players.findIndex((player) => player.id === userId) + 1;
-};
-
-const calculateGwAction = (gameweek: Gameweek): string => {
-  return isOpen(gameweek) ? 'Create bet' : 'Locked';
-};
-
-const isOpen = (gameweek: Gameweek): boolean => {
-  const now = new Date();
-  const opens = new Date(gameweek.opens);
-  const closes = new Date(gameweek.closes);
-
-  return opens < now && now < closes;
-};
+import Scoreboard from '../../../components/leagueId/Scoreboard';
+import {
+  calculateGwAction,
+  calculateTimeUntilGW,
+  calculateYourPlace,
+  isOpen,
+} from '../../../utils/leagueUtils';
 
 const LeagueView = () => {
   const router = useRouter();
@@ -223,25 +109,11 @@ const LeagueView = () => {
 };
 
 const styles = StyleSheet.create({
-  scoreboard: {
-    backgroundColor: colors.charcoal[2],
-    padding: 2,
-  },
   betWrapper: {
     padding: 12,
     display: 'flex',
     gap: 4,
     alignItems: 'center',
-  },
-  scoreboardScrollWrapper: {
-    display: 'flex',
-    gap: 8,
-  },
-  scoreboardHeader: {
-    padding: 4,
-  },
-  centeredText: {
-    textAlign: 'center',
   },
 
   statsWrapper: {
@@ -256,12 +128,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
     padding: 8,
-  },
-
-  textWrapper: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
   },
 });
